Create custom hook to handle setting bounds to feature collection
