+++ moped-editor/src/views/projects/projectView/ProjectSummaryMap.js
@@ -1,10 +1,5 @@
-import React, { useState, useRef, useEffect } from "react";
-import ReactMapGL, {
-  Layer,
-  NavigationControl,
-  Source,
-  WebMercatorViewport,
-} from "react-map-gl";
+import React, { useRef } from "react";
+import ReactMapGL, { Layer, NavigationControl, Source } from "react-map-gl";
 import { Box, Button, makeStyles } from "@material-ui/core";
 import { EditLocation as EditLocationIcon } from "@material-ui/icons";
 import "mapbox-gl/dist/mapbox-gl.css";
@@ -13,13 +8,12 @@ import "react-map-gl-geocoder/dist/mapbox-gl-geocoder.css";
 import {
   createProjectViewLayerConfig,
   MAPBOX_TOKEN,
-  mapConfig,
   mapStyles,
   renderTooltip,
   renderFeatureCount,
   sumFeaturesSelected,
   useHoverLayer,
-  createZoomBbox,
+  useFeatureCollectionToFitBounds,
 } from "../../../utils/mapHelpers";
 
 const useStyles = makeStyles({
@@ -50,46 +44,17 @@ const ProjectSummaryMap = ({
   const mapRef = useRef();
   const featureCount = sumFeaturesSelected(selectedLayerIds);
 
-  const [viewport, setViewport] = useState(mapConfig.mapInit);
   const { handleLayerHover, featureId, hoveredCoords } = useHoverLayer();
+  const [viewport, setViewport] = useFeatureCollectionToFitBounds(
+    mapRef,
+    projectExtentGeoJSON
+  );
 
   /**
    * Updates viewport on zoom, scroll, and other events
    * @param {Object} viewport - Mapbox object that stores properties of the map view
    */
   const handleViewportChange = viewport => setViewport(viewport);
-
-  useEffect(() => {
-    const currentMap = mapRef.current;
-    const mapBounds = createZoomBbox(projectExtentGeoJSON);
-
-    /**
-     * Takes the existing viewport and transforms it to fit the project's features
-     * @param {Object} viewport - Describes the map view
-     * @return {Object} Viewport object with updated attributes based on project's features
-     */
-    const fitViewportToBounds = viewport => {
-      const featureViewport = new WebMercatorViewport({
-        viewport,
-        width: currentMap._width,
-        height: currentMap._height,
-      });
-      const newViewport = featureViewport.fitBounds(mapBounds, {
-        padding: 100,
-      });
-
-      const { longitude, latitude, zoom } = newViewport;
-
-      return {
-        ...viewport,
-        longitude,
-        latitude,
-        zoom,
-      };
-    };
-
-    setViewport(prevViewport => fitViewportToBounds(prevViewport));
-  }, [projectExtentGeoJSON]);
 
   return (
     <Box>
